+++ oneflow/graph/node.h
@@ -41,7 +41,7 @@ class Edge {
   }
   virtual ~Edge() = default;
 
-  int32_t edge_id() const { return edge_id_; }
+  int64_t edge_id() const { return edge_id_; }
   std::string edge_id_str() const { return std::to_string(edge_id_); }
 
   NodeType* src_node() const { return src_node_; }
@@ -53,7 +53,7 @@ class Edge {
                                           NodeType* dst_node);
   friend void DisConnect<EdgeType>(EdgeType* edge);
   
-  int32_t edge_id_;
+  int64_t edge_id_;
   
   NodeType* src_node_;
   NodeType* dst_node_;
@@ -69,7 +69,7 @@ class Node {
   }
   virtual ~Node() = default;
 
-  int32_t node_id() const { return node_id_; }
+  int64_t node_id() const { return node_id_; }
   std::string node_id_str() const { return std::to_string(node_id_); }
   EdgeType* SoleInEdge() const {
     CHECK_EQ(in_edges_.size(), 1);
@@ -102,7 +102,7 @@ class Node {
                                           NodeType* dst_node);
   friend void DisConnect<EdgeType>(EdgeType* edge);
 
-  int32_t node_id_;
+  int64_t node_id_;
 
   std::unordered_set<EdgeType*> in_edges_;
   std::unordered_set<EdgeType*> out_edges_;
